pyproject.toml
[project]
name = "medical-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.1.1",
    "flask-cors>=6.0.1",
    "google-ai-generativelanguage>=0.6.15",
    "google-generativeai>=0.8.5",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-experimental>=0.3.4",
    "langchain-google-genai>=2.0.10",
    "langchain-openai>=0.3.28",
    "langchain-pinecone>=0.2.8",
    "pinecone[grpc]>=7.3.0",
    "pypdf>=5.8.0",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.0.0",
    "streamlit>=1.46.1",
]
